agents: MCTS: refine edge addon update logic

+++ agents/include/MCTS/policy/CreditPolicy.h
@@ -12,20 +12,19 @@ namespace mcts
 			// * If the first player has 100% to loss, the credit should be 0.0
 			// * If the first player has 50% to win, the credit should be 0.5
 			// * The credit value should be in range [0.0, 1.0]
-
-			static double GetCredit(engine::view::Board const& board, engine::Result result) {
+			static float GetCredit(engine::view::Board const& board, engine::Result result) {
 				auto side = board.GetViewSide();
 				bool winning = WinOrLoss(side, result);
 
-				double score = 0.0;
-				if (winning) score = 1.0;
-				else score = -1.0;
+				float score = 0.0f;
+				if (winning) score = 1.0f;
+				else score = -1.0f;
 
-				static constexpr double kMinScore = 0.5;
-				score = (score * (1.0 - kMinScore)) + kMinScore;
+				static constexpr float kMinScore = 0.5f;
+				score = (score * (1.0f - kMinScore)) + kMinScore;
 				
 				// scale from [-1, 1] to [0, 1]
-				score = (score + 1.0) * 0.5;
+				score = (score + 1.0f) * 0.5f;
 				return score;
 			}
